Updated releng/autorelease Project: bgpcep  4a3e1595445dad87ab6f6990961639a0046dcd0f
Bug 4488: Metadata not available for modification NodeModification

To resovle tis race condition problem, modify EffectiveRibInWriter to
work on per peer basis instead of a global per RIB instance.
The EffectiveRibInWriter is now listening on paricular peer's AdjRibIn tables.
The writer is started when a session goes up and closed when the session goes down.
Peer role is not tracked via data changes rather directly set at creation.

The possitive side effect is better error recovery in a case of failure in
EffectiveRibInWriter.
Disadvantage of this change is tighter coupling between components
(peer, eff-rib-in writer and policy tracker).

Change-Id: I650bae8638a2a4818bdd925a5f4ca2274b79d5c6
Signed-off-by: Milos Fabian <milfabia@cisco.com>