Updated releng/autorelease Project: ovsdb  808787affcd5586a27b6bb3ae779d6f3a7f4b900
BUG-5507: ignore schema-related AutoAttach errors

When initialising OvsdbAutoAttachedRemovedCommand and
OvsdbAutoAttachUpdateCommand, schemas without support for autoattach
cause IllegalArgumentException to be thrown. This is in all likelihood
the only scenario which can lead to this, so for now ignore the
exceptions.

A proper fix would involve checking the schema as done in TyperUtils.
Once this is done these constructors will be updated to use that
instead of ignoring IllegalArgumentException.

Change-Id: Idd284ab8acdc0886ef11dae6ecfb410e487ca312
Signed-off-by: Stephen Kitt <skitt@redhat.com>