Updated git submodules
Project: mdsal master a034bf9bbc7abef6a749b9ed658d74a90cea1a0e

Do not use compile dependency for embedded code

We are embedding xtend and its dependencies into our jar and using
compile dependency means karaf plugin will download the jars as it
will see it as a runtime dependency.

Degrade the dependency to scope=provided, as we will provide the
dependencies by our own jar. This breaks maven-xtend-plugin, which
needs guava<15 and relies on it being pulled in transitively. Fix that
by adding an explitic dependency in plugin declaration.

Change-Id: Idb358bca4ce8eab7179d6867e603dd6b2306e732
Signed-off-by: Robert Varga <rovarga@cisco.com>


Do not depend on java-api-generator

The api generator is not part of features, do not specify it as a dependency.
Same is true for the maven plugin.

Change-Id: I6ca54e65e6b98662b087c95d887acf448b952731
Signed-off-by: Robert Varga <rovarga@cisco.com>