Update patch set 1
Patch Set 1: Code-Review-1

(15 comments)

The thing I care about fixing most is the copyright headers. Please fix it to be an actual entity such as Huawei or An Ho.

The rest are pretty minor but we should probably fix the trailing whitespace as well as make sure tabs|spaces are consistent throughout all the files.

Patch-set: 1
Label: Code-Review=-1