Updated git submodules
Project: netvirt master f145ef664ce4d617ac836d1f95ba37e4bfb50ccf

Merge "Remove the unused shared network validation check as its a bit misleading"

Remove the unused shared network validation check as its a bit misleading

ODL Neutron northbound is now MDSAL based (from Beryllium).
The canCreateNetwork/canUpdateNetwork checks are no longer
invoked in NetworkHandler of netvirt.

Observations with stable/Beryllium and master branch.
-----------------------------------------------------
In stable/beryllium and master branches an admin is able
to create a shared network. Two tenants are able to spawn
VMs (on the shared network), which successfully acquire
dhcp-ip and are able to communicate with the VM belonging
to the other tenant. Verified this use-case on a multi-node
setup.

It appears like shared networks are indeed supported. So, we
can remove this un-used validation as its misleading.

Observations with stable/Lithium SR4 release.
---------------------------------------------
An admin is unable to create a shared tenant network but shared
external networks are supported. Before removing the validation
for Lithium branch we have to see if there was any specific
reason why the check was inplace.

Change-Id: I62ab980110c58f07a9e161431b64a64590c1e410
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>