Updated git submodules
Project: aaa master c7e356e203342a492a3311a6ffacc6433113a73c

Depend on odlparent version of logback

Downstream dependencies should centralize dependency management
in odlparent.  This change technically temporarily downgrades
logback to 1.1.3, which is the one included with odlparent.  This
way, when logback is upgraded in odlparent, we get it for free.

Change-Id: Ibce99e34bc65db678390d37169c5cd924113f389
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>