Updated git submodules
Project: lispflowmapping master d37a97737588593a2222c94011477c65de796e09

Remove maskLength from MappingRecord

The maskLength field was deprecated in Beryllium, when we switched to
the IETF LISP address type model, and its value is now integral part if
the Eid field, which can be a prefix. We can now remove it, to avoid
further confusion.

Change-Id: I7f464bc157065b79bab6110eecf0c71c16aa125b
Signed-off-by: Lorand Jakab <lojakab@cisco.com>