Updated git submodules
Project: openflowjava master 693ca12245dd55cad42ef3da61f409a5c678eb37

Add SwitchConnectionProviderFactory

Added a SwitchConnectionProviderFactory interface to the protocol SPI
bundle and a SwitchConnectionProviderFactoryImpl to the impl bundle to
create SwitchConnectionProvider instances given a SwitchConnectionConfig.
The SwitchConnectionProviderFactoryImpl is instantiated and advertised as an
OSGi service via blueprint. This allows clients to create new
SwitchConnectionProvider instances while hiding implementation details.
This is equivalent to creating instances via the config yang module.

The SwitchConnectionConfig and related classes are generated via a new
yang file that is equivalent to the config elements defined in the
openflow-switch-connection-provider-impl config yang. This new yang
model will be stored and retrieved via the data store.

Change-Id: I511c12644d5d54cd99c6b1afbf4c078a0cecee8e
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>