Updated git submodules
Project: openflowplugin master 346070695dff1f6b3afe7a79437c04c5523159ca

Add the openflowplugin-blueprint-config bundles to the features xml

This turns on the blueprint wiring for the openflow helium plugin

Change-Id: I5725495f3a40e636bc35cb1a9281c593ad072840
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>


Add blueprint wiring for the lithium OpenflowPluginProvider

Change-Id: I95d721e85f0116b7d83c6b2db475ddfc27aa70c9
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>


Add blueprint wiring for the helium inventory manager

Change-Id: Ie6685cdf636c6f707943460bc6dbcac3483442c8
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>


Add blueprint wiring for the helium StatisticsManager

Change-Id: I81ac14b8ea9e4d76d3088ae512a52cfc758c8688
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>


Add blueprint wiring for the helium OpenflowPluginProvider

Added a new bundle, openflowplugin-blueprint-config-he, containing the
blueprint XML that instantiates the SwitchConnectionProvider
and OpenflowPluginProvider instances for the helium plugin. It uses
the new SwitchConnectionProviderFactory interface added to the
openflowjava SPI.

This replaces the ConfigurableOpenFlowProviderModule which is now
deprecated but remains for backwards compatibility. The createInstance
method now obtains the OpenflowPluginProvider created via blueprint from
the OSGi registry and returns a proxy instance that delegates to the
actual instance but overrides the close method appropriately.

Patch 3 : Resolved conflict locally

Change-Id: I315444847f5a95d905a5e6adc602445c1101da5b
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>