Updated git submodules
Project: ovsdb stable/beryllium 568dd2a36c6a92200a7a1c91416df946946469b6

Merge "BUG 5885 - OVSDB plugin failure to update passive ovsdb nodes" into stable/beryllium

BUG 5885 - OVSDB plugin failure to update passive ovsdb nodes

Updates the ovsdb southbound so that configuration to ovsdb
nodes which have connected to the plugin passively can be
updated (e.g. such as qos or queue entries) without requiring
connection-info to be supplied in the configuration.
Similar to what already works when adding bridges or
termination points to passively connected ovsdb nodes.

The onDataChanged() method executes the following
sequence of operations on changes to the Ovsdb data tree:
- connect(changes)
- updateConnections(changes)
- updateData(changes)
- disconnect(changes)

When an OVSDB node has connected passively, or more
generally, when the changes to the OVSDB node do not
include 'connection-info', it is not necessary to
attempt to execute connection related operations for
the specified host.  This patch  checks for the absence
of the connection-info and skips trying to connect or
update a connection in the connect() and updateConnections()
methods.

v2 - add in data tree change patch from master branch
     now that data tree change listener is merged.

Change-Id: I8238711076a74cfa06422e5cacb127ba95fd23b7
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>